backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose
passlib[bcrypt]
python-multipart
